Drop recipes also picked up from `TestNGToAssertJRulesRecipes`
