+++ src/main/java/org/openrewrite/java/testing/testng/TestNgAssertNotEqualsToAssertThat.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java.testing.assertj.testng;
+package org.openrewrite.java.testing.testng;
 
 import org.openrewrite.ExecutionContext;
 import org.openrewrite.Preconditions;
@@ -31,19 +31,19 @@ import org.openrewrite.java.tree.TypeUtils;
 
 import java.util.List;
 
-public class TestNgAssertEqualsToAssertThat extends Recipe {
+public class TestNgAssertNotEqualsToAssertThat extends Recipe {
 
     @Override
     public String getDisplayName() {
-        return "TestNG `assertEquals` to AssertJ";
+        return "TestNG `assertNotEquals` to AssertJ";
     }
 
     @Override
     public String getDescription() {
-        return "Convert TestNG-style `assertEquals()` to AssertJ's `assertThat().isEqualTo()`.";
+        return "Convert TestNG-style `assertNotEquals()` to AssertJ's `assertThat().isNotEqualTo()`.";
     }
 
-    private static final MethodMatcher TESTNG_ASSERT_METHOD = new MethodMatcher("org.testng.Assert assertEquals(..)");
+    private static final MethodMatcher TESTNG_ASSERT_METHOD = new MethodMatcher("org.testng.Assert assertNotEquals(..)");
 
     @Override
     public TreeVisitor<?, ExecutionContext> getVisitor() {
@@ -55,30 +55,32 @@ public class TestNgAssertEqualsToAssertThat extends Recipe {
                 }
 
                 List<Expression> args = method.getArguments();
+
                 Expression expected = args.get(1);
                 Expression actual = args.get(0);
 
-                //always add the import (even if not referenced)
-                maybeAddImport("org.assertj.core.api.Assertions", "assertThat", false);
-
-                // Remove import for "org.testng.Assert" if no longer used.
-                maybeRemoveImport("org.testng.Assert");
-
                 if (args.size() == 2) {
-                    return JavaTemplate.builder("assertThat(#{any()}).isEqualTo(#{any()});")
+                    method = JavaTemplate.builder("assertThat(#{any()}).isNotEqualTo(#{any()});")
                             .staticImports("org.assertj.core.api.Assertions.assertThat")
                             .javaParser(JavaParser.fromJavaVersion()
                                     .classpathFromResources(ctx, "assertj-core-3.24"))
                             .build()
-                            .apply(getCursor(), method.getCoordinates().replace(), actual, expected);
+                            .apply(
+                                    getCursor(),
+                                    method.getCoordinates().replace(),
+                                    actual,
+                                    expected
+                            );
                 } else if (args.size() == 3 && !isFloatingPointType(args.get(2))) {
                     Expression message = args.get(2);
+
                     JavaTemplate.Builder template = TypeUtils.isString(message.getType()) ?
-                            JavaTemplate.builder("assertThat(#{any()}).as(#{any(String)}).isEqualTo(#{any()});") :
-                            JavaTemplate.builder("assertThat(#{any()}).as(#{any(java.util.function.Supplier)}).isEqualTo(#{any()});");
-                    return template
+                            JavaTemplate.builder("assertThat(#{any()}).as(#{any(String)}).isNotEqualTo(#{any()});") :
+                            JavaTemplate.builder("assertThat(#{any()}).as(#{any(java.util.function.Supplier)}).isNotEqualTo(#{any()});");
+
+
+                    method = template
                             .staticImports("org.assertj.core.api.Assertions.assertThat")
-                            .imports("java.util.function.Supplier")
                             .javaParser(JavaParser.fromJavaVersion()
                                     .classpathFromResources(ctx, "assertj-core-3.24"))
                             .build()
@@ -90,43 +92,53 @@ public class TestNgAssertEqualsToAssertThat extends Recipe {
                                     expected
                             );
                 } else if (args.size() == 3) {
-                    //always add the import (even if not referenced)
-                    maybeAddImport("org.assertj.core.api.Assertions", "within", false);
-                    return JavaTemplate.builder("assertThat(#{any()}).isCloseTo(#{any()}, within(#{any()}));")
+                    method = JavaTemplate.builder("assertThat(#{any()}).isNotCloseTo(#{any()}, within(#{any()}));")
                             .staticImports("org.assertj.core.api.Assertions.assertThat", "org.assertj.core.api.Assertions.within")
                             .javaParser(JavaParser.fromJavaVersion()
                                     .classpathFromResources(ctx, "assertj-core-3.24"))
                             .build()
-                            .apply(getCursor(), method.getCoordinates().replace(), actual, expected, args.get(2));
+                            .apply(
+                                    getCursor(),
+                                    method.getCoordinates().replace(),
+                                    actual,
+                                    expected,
+                                    args.get(2)
+                            );
+                    maybeAddImport("org.assertj.core.api.Assertions", "within", false);
+                } else {
+                    Expression message = args.get(3);
 
+                    JavaTemplate.Builder template = TypeUtils.isString(message.getType()) ?
+                            JavaTemplate.builder("assertThat(#{any()}).as(#{any(String)}).isNotCloseTo(#{any()}, within(#{any()}));") :
+                            JavaTemplate.builder("assertThat(#{any()}).as(#{any(java.util.function.Supplier)}).isNotCloseTo(#{any()}, within(#{any()}));");
+
+                    method = template
+                            .staticImports("org.assertj.core.api.Assertions.assertThat", "org.assertj.core.api.Assertions.within")
+                            .javaParser(JavaParser.fromJavaVersion()
+                                    .classpathFromResources(ctx, "assertj-core-3.24"))
+                            .build()
+                            .apply(
+                                    getCursor(),
+                                    method.getCoordinates().replace(),
+                                    actual,
+                                    message,
+                                    expected,
+                                    args.get(2)
+                            );
+
+                    maybeAddImport("org.assertj.core.api.Assertions", "within", false);
                 }
 
-                // The assertEquals is using a floating point with a delta argument and a message.
-                Expression message = args.get(3);
+                //Make sure there is a static import for "org.assertj.core.api.Assertions.assertThat" (even if not referenced)
+                maybeAddImport("org.assertj.core.api.Assertions", "assertThat", false);
 
-                //always add the import (even if not referenced)
-                maybeAddImport("org.assertj.core.api.Assertions", "within", false);
-                JavaTemplate.Builder template = TypeUtils.isString(message.getType()) ?
-                        JavaTemplate.builder("assertThat(#{any()}).as(#{any(String)}).isCloseTo(#{any()}, within(#{any()}));") :
-                        JavaTemplate.builder("assertThat(#{any()}).as(#{any(java.util.function.Supplier)}).isCloseTo(#{any()}, within(#{any()}));");
-                return template
-                        .staticImports("org.assertj.core.api.Assertions.assertThat", "org.assertj.core.api.Assertions.within")
-                        .imports("java.util.function.Supplier")
-                        .javaParser(JavaParser.fromJavaVersion()
-                                .classpathFromResources(ctx, "assertj-core-3.24"))
-                        .build()
-                        .apply(
-                                getCursor(),
-                                method.getCoordinates().replace(),
-                                actual,
-                                message,
-                                expected,
-                                args.get(2)
-                        );
+                // Remove import for "org.testng.Assert" if no longer used.
+                maybeRemoveImport("org.testng.Assert");
+
+                return method;
             }
 
             private boolean isFloatingPointType(Expression expression) {
-
                 JavaType.FullyQualified fullyQualified = TypeUtils.asFullyQualified(expression.getType());
                 if (fullyQualified != null) {
                     String typeName = fullyQualified.getFullyQualifiedName();
